[OU-IMP] sale: review followup
