+++ openupgrade_scripts/scripts/sale/tests/test_sale_migration.py
@@ -1,0 +1,10 @@
+from odoo.tests import TransactionCase
+
+from odoo.addons.openupgrade_framework import openupgrade_test
+
+
+@openupgrade_test
+class TestSaleMigration(TransactionCase):
+    def test_sale_order_state(self):
+        self.assertEqual(self.env.ref("sale.sale_order_18").state, "sale")
+        self.assertFalse(self.env.ref("sale.sale_order_18").locked)
